+++ src/content/index.ts
@@ -9,6 +9,7 @@ import { TUICI18N } from "./i18n.ts";
 import { applySystemCss, addCssElement, applyDataCss, applyCustomIcon } from "./applyCSS.ts";
 import { isSafemode, runSafemode } from "./safemode.ts";
 import { startTluiObserver } from "@shared/tlui/observer.ts";
+import { twitterIcon } from "./modules/observer/functions.ts";
 
 (async () => {
     await TUICI18N.fetch();
@@ -34,7 +35,7 @@ import { startTluiObserver } from "@shared/tlui/observer.ts";
     applyDataCss();
     applyCustomIcon();
     if (document.querySelector(`#placeholder > svg`)) {
-        TUICObserver.functions.twitterIcon(document.querySelector(`#placeholder > svg:not(.NOT_TUIC_DISPNONE):not(.TUIC_DISPNONE`), document.querySelector(`#placeholder`));
+        twitterIcon(document.querySelector(`#placeholder > svg:not(.NOT_TUIC_DISPNONE):not(.TUIC_DISPNONE`), document.querySelector(`#placeholder`));
     }
 
     chrome.runtime.sendMessage({
